New skeleton mesh in Unreal
Added the new correctly rotated skeleton mesh into the blueprint in unreal. Made them have physics. The move function is disabled, but they are moving on their own, i have no idea why. Also changed the collision of the tree with leaves, so that it is possible to walk past it.